move clap structs to cli module
