+++ host/src/lib.rs
@@ -19,6 +19,8 @@ use std::{
 
 use risc0_zkvm::Receipt;
 
+pub mod cli;
+
 pub fn cache_file_path(cache_path: &Path, network: &str, block_no: u64, ext: &str) -> PathBuf {
     cache_path
         .join(network)
